refactor:widget파일을 두개의 파일로 분리
